move component styles to separate folder in styles/
