+++ src/components/IndexPage/builds.js
@@ -1,9 +1,10 @@
 import React from "react"
 import PropTypes from "prop-types"
-import { HeaderLineBelow } from "styles/Header"
 import styled from "styled-components"
+
 import Project from "./project.js"
-import Section from "styles/Section"
+import { HeaderLineBelow } from "styles/components/Header"
+import Section from "styles/components/Section"
 
 const ProjectContainer = styled.div`
   display: flex;
