+++ src/components/footer.js
@@ -1,6 +1,6 @@
 import React from "react"
 import PropTypes from "prop-types"
-import Section from "../styles/Section"
+import Section from "styles/components/Section"
 import styled from "styled-components"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
 import {
